+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Bug fix
 *******
 
 * Fix document broken by multiple references to a same footnote.
+* Fix documentation build error by broken bookmark in Windows.
 
 Release 1.1.4 (2019-09-25)
 --------------------------